+++ src/ripper.rb
@@ -4,7 +4,8 @@ require 'json'
 require 'ripper'
 
 class RipperJS < Ripper::SexpBuilder
-  attr_reader :block_comments, :inline_comment, :current_embdoc
+  attr_reader :block_comments, :inline_comment
+  attr_reader :current_embdoc, :current_heredoc
   attr_reader :last_sexp
 
   def initialize(*args)
@@ -16,7 +17,9 @@ class RipperJS < Ripper::SexpBuilder
 
     @block_comments = []
     @inline_comment = nil
+
     @current_embdoc = nil
+    @current_heredoc = nil
 
     @last_sexp = nil
   end
@@ -117,6 +120,20 @@ class RipperJS < Ripper::SexpBuilder
     current_embdoc[:body] << comment.chomp
     block_comments << current_embdoc
     @current_embdoc = nil
+  end
+
+  def on_heredoc_beg(beginning)
+    @current_heredoc = { type: :heredoc, beginning: beginning, start: lineno, end: lineno }
+  end
+
+  def on_heredoc_end(ending)
+    current_heredoc.merge!(ending: ending.chomp, end: lineno)
+  end
+
+  def on_heredoc_dedent(string, _width)
+    string.merge!(current_heredoc.slice(:type, :beginning, :ending, :start, :end)).tap do
+      @current_heredoc = nil
+    end
   end
 
   def on_method_add_block(*body)