+++ test/strings.rb
@@ -30,4 +30,16 @@
 
 %x[super_super_super_super_super_super_super_super_super_super_super_super_su_long]
 
+<<~HERE
+  This is a squiggly heredoc!
+HERE
+
+<<~HERE
+    This is another squiggly heredoc!
+HERE
+
+abc = <<~HERE
+  This is a squiggly heredoc on an assign!
+HERE
+
 # rubocop:enable Lint/Void